build-transaction-cron/requirements_cron.txt
ibm_db>=3.1.0
python-dotenv>=1.0.0